### 2020-08-05 #2549 @whedon
[PRE REVIEW]: qgs: A flexible Python framework of reduced-order multiscale climate models
**Submitting author:** @jodemaey (<a href="http://orcid.org/0000-0002-5098-404X">Jonathan Demaeyer</a>)
**Repository:** <a href="https://github.com/Climdyn/qgs" target ="_blank">https://github.com/Climdyn/qgs</a>
**Version:** v0.2.0
