### 2020-08-05 #2549 @danielskatz
👋 @jodemaey

 - please check the DOIs as mentioned in a previous comment - whedon is not always right, but makes suggestions.  If you make changes in your bib, enter the whedon commands I used above to regenerate the PDF and recheck the references
 - In the paper, in the state of the field section, please explain how your work is different/related
