### 2020-08-05 #2549 @harpolea
@jodemaey do you have any suggestions for potential reviewers?  If so, then please mention them here in this thread (without tagging them with an @). This [list of people ](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
