### 2020-08-05 #2549 @harpolea
@jodemaey thanks!

:wave: @eviatarbach & @rabernat, would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
