### 2020-08-06 #2549 @jodemaey
>     * please check the DOIs as mentioned in a previous comment - whedon is not always right, but makes suggestions.  If you make changes in your bib, enter the whedon commands I used above to regenerate the PDF and recheck the references
>

Here whedon is not right about the missing DOIs. The one for the reference to the Lorenz 96 paper points in fact to a re-edition of it in 2006 and I want to point to the original 1996 paper which has no doi. The one about Isca is also wrong. To my knowledge, the Isca model doesn't have a DOI yet.

About the DOI that whedon claims to be wrong, if I click on all of them they lead me to the correct articles. All of them contains semicolons and weird characters so I guess that whedon has a problem with that.
