+++ documentation/notes/joss_paper/joss.bib
@@ -127,6 +127,18 @@
   year={1996}
 }
 
+@inbook{L2006,
+  place={Cambridge},
+  title={Predictability – a problem partly solved},
+  doi={10.1017/CBO9780511617652.004},
+  booktitle={Predictability of Weather and Climate},
+  publisher={Cambridge University Press},
+  author={Lorenz, Edward N.},
+  editor={Palmer, Tim and Hagedorn, RenateEditors},
+  year={2006},
+  pages={40–58}
+}
+
 @article{L84,
   title={Irregularity: A fundamental property of the atmosphere},
   author={Lorenz, Edward N},
@@ -263,3 +275,13 @@
   doi={10.1007/s00382-020-05313-3}
 }
 
+@article{Vetal2018,
+  author={Vallis, G. K. and Colyer, G. and Geen, R. and Gerber, E. and Jucker, M. and Maher, P. and Paterson, A. and Pietschnig, M. and Penn, J. and Thomson, S. I.},
+  title={Isca, v1.0: a framework for the global modelling of the atmospheres of Earth and other planets at varying levels of complexity},
+  journal={Geoscientific Model Development},
+  volume={11},
+  year={2018},
+  number={3},
+  pages={843--859},
+  doi={10.5194/gmd-11-843-2018}
+}
