### 2020-08-24 #2549 @harpolea
Thanks!

:wave: @sadielbartholomew, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
