### 2020-08-24 #2549 @rabernat
Hi all--sorry for the slow response. I have been on vacation for most of August.

I am not able to review this submission right now. Some great alternatives might be @mfjansen, @francispoulin, @navidcy, @pittwolfe, and @mbueti, all of whom have worked on [pyqg](https://github.com/pyqg/pyqg).
