### 2020-08-24 #2549 @sadielbartholomew
> wave @sadielbartholomew, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi, thanks for the request and yes I am happy to review this, assuming you are happy to wait one week or so for me to start reviewing since I am currently a JOSS reviewer for another library and will need to devote a little more time to that review before I should accept another.
