### 2020-08-24 #2549 @harpolea
@sadielbartholomew great, thanks! And yes, that's totally fine – we typically ask that reviews are completed within 6 weeks or so, so that should give you plenty of time to finish up with the other review first!
