### 2020-08-31 #2597 @eviatarbach
qgs provides a modular implementation of a quasi-geostrophic atmospheric model. Having previously used the MAOOAM model in research, I was impressed with its ease-of-use, highly readable code, the elegance of the tensor-based implementation, and built-in tangent linear and adjoint models. qgs builds on this by adding two different land components, and making it easy to switch between ocean and land coupling. qgs is also written entirely in Python, as opposed to MAOOAM that had Fortran and Lua implementations also. I commend the authors for their excellent contribution and hope that qgs will become a widely used tool in the modelling hierarchy.

Suggested changes:
1. There are a few missing parts of the documentation, as per the review checklist. The statement of need is provided in the software paper, but it has to be in the documentation as well.
> - A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> - Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
2. I would like to see the software paper mention the availability of the tangent linear and adjoint models, as this is a major advantage for data assimilation research, as well as for, e.g., computing the Lyapunov spectrum. Adding an example of the usage of the tangent linear or adjoint models to the documentation would be helpful.
3. This implementation is written in Python instead of Fortran, but uses Numba and sparse matrices to accelerate the computations. I would thus like to see a performance comparison to the Fortran version of MAOOAM, either in the documentation or the software paper (see the performance item on the checklist).

Minor issues:
- "Simpy" should be Sympy in the forthcoming developments section.
- The link to the notebooks folder works in the GitHub README, but is broken in the readthedocs documentation. Also, the link to the main folder just goes to the main page of the documentation.
- "montain" -> "mountain" in the Reinhold and Pierrehumbert (1982) example page
