### 2020-10-05 #2597 @kthyng
Hi @jodemaey! I am the rotating associate editor in chief this week and am going through stale submissions. I see that one of your reviewers put forth some comments above but you haven't responded yet. Will you be able to get to this soon? If not, that is ok, but we should in that case pause this submission. I recommend that we pause this if we haven't heard back from the author within a week or so.
