### 2020-10-05 #2597 @sadielbartholomew
Regarding your comment @kthyng, just to clarify that I was waiting on the submitting author's response to the first reviewers' comments, as in https://github.com/openjournals/joss-reviews/issues/2597#issuecomment-683972381, before starting my review, in case there were any significant changes made to the codebase or paper based on that feedback.

However, as you state it has been a while now, a good month, since that feedback without response. Would you recommend I start my review regardless, now? Or wait considering this could be considered 'stale'? I am just wary we are approaching six weeks since the review was opened and it is recommended I complete my review within that timeframe.
