### 2020-10-05 #2597 @jodemaey
Hi @kthyng and @sadielbartholomew,

There was a misunderstanding here. I was waiting for the second review to address all the comments at once.

In any case I'm not able to work on this until the end of October. At that time I will work again fully on qgs and will address the comments and suggestions of @eviatarbach .
