### 2020-10-05 #2597 @sadielbartholomew
Sorry that there was this misunderstanding, @jodemaey.

If you are unable to make changes until November it sounds like it would be best for me to complete my review by then so that you can know what else, if anything, might need to be addressed. I will ensure I have my review completed by then, and try to do it much sooner anyhow, now that I know about the situation. Thanks.
