### 2020-10-14 bac1d3c
Corrected minor issues raised by Eviatar Bach (JOSS reviewer #1).
