+++ documentation/generate_examples.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-cd source/files/examples
+cd source/files/examples || exit
 
 for file in ./*.ipynb; do
     jupyter nbconvert --to rst "$file"
