+++ documentation/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2020, Jonathan Demaeyer and Lesley De Cruz'
 author = 'Jonathan Demaeyer and Lesley De Cruz'
 
 # The full version, including alpha/beta/rc tags
-release = 'v0.1.0'
+release = 'v0.2.0'
 version = release
 
 
