### 2020-10-14 f87022c
Addressed the first suggestion of Eviatar Bach (JOSS reviewer #1)
