+++ README.md
@@ -128,6 +128,27 @@ Forthcoming developments
     + Active advection
     + True quasi-geostrophic ocean when using ocean model version
     + Salinity in the ocean
+    
+Contributing to qgs
+-------------------
+
+If you want to contribute actively to the roadmap detailed above, please contact directly the authors.
+
+In addition, if you have made changes that you think will be useful to others, please feel free to suggest these as a pull request on the [qgs Github repository](https://github.com/Climdyn/qgs).
+
+A review of your pull request will follow with possibly suggestions of changes before merging it in the master branch.
+Please consider the following guidelines before submitting:
+* Before submitting a pull request, double check that the branch to be merged contains only changes you wish to add to the master branch. This will save time in reviewing the code.
+* For any changes to the core model files, please run the tests found in the folder [model_test](./model_test) to ensure that the model tensors are still valid.
+* For substantial additions of code, including a test case in the folder [model_test](./model_test) is recommended.
+* Please do not make changes to existing test cases, these are here for trip-testing as well as user guidance.
+* Please document the new functionalities in the documentation. Code addition without documentation addition will not be accepted. 
+The documentation is done with [sphinx](https://www.sphinx-doc.org/en/master/) and follows the Numpy conventions. Please take a look to the actual code to get an idea about how to document the code.
+* If your addition can be considered as a tool not directly related to the core of the model, please develop it in the toolbox folder.
+* The team presently maintaining qgs is not working full-time on it, so please be patient as the review of the submission may take some times.
+
+For more information about git, Github and the pull request framework, a good source of information is the [contributing guide](https://mitgcm.readthedocs.io/en/latest/contributing/contributing.html) of the [MITgcm](https://github.com/MITgcm/MITgcm).
+
 
 Other atmospheric models in Python
 ----------------------------------
