+++ documentation/source/files/model/ref.bib
@@ -417,3 +417,15 @@ year={2014}
   pages={1--18},
   year={2000}
 }
+
+@article{OB1989,
+  author = {Enda O'Brien and Lee Branscome},
+  title = {Minimal modeling of the extratropical general circulation},
+  journal = {Tellus A: Dynamic Meteorology and Oceanography},
+  volume = {41},
+  number = {4},
+  pages = {292-307},
+  year  = {1989},
+  publisher = {Taylor & Francis},
+  url = {https://doi.org/10.3402/tellusa.v41i4.11842}
+}
