### 2020-10-14 cee79b7
The tangent linear model is now mentioned in the paper and an example was added to the documentation
