+++ documentation/execute_examples.sh
@@ -3,6 +3,6 @@
 cd source/files/examples || exit
 
 for file in ./*.ipynb; do
-    jupyter nbconvert --execute --inplace "$file"
+    jupyter nbconvert --execute --inplace --ExecutePreprocessor.timeout=600 "$file"
 done
 
