+++ documentation/notes/joss_paper/joss.bib
@@ -297,3 +297,10 @@ year  = {1989},
 publisher = {Taylor & Francis},
 doi = {10.3402/tellusa.v41i4.11842}
 }
+
+@book{K2003,
+  title={Atmospheric modeling, data assimilation and predictability},
+  author={Kalnay, Eugenia},
+  year={2003},
+  publisher={Cambridge university press}
+}
