### 2020-10-20 58cb5d0
Added 6x6 model version tests
