### 2020-10-26 dfaeed9
First draft of the performance section asked by Reviewer #1
