+++ documentation/notes/joss_paper/joss.bib
@@ -287,15 +287,15 @@
 }
 
 @article{OB1989,
-author = {Enda O'Brien and Lee Branscome},
-title = {Minimal modeling of the extratropical general circulation},
-journal = {Tellus A: Dynamic Meteorology and Oceanography},
-volume = {41},
-number = {4},
-pages = {292-307},
-year  = {1989},
-publisher = {Taylor & Francis},
-doi = {10.3402/tellusa.v41i4.11842}
+  author = {Enda O'Brien and Lee Branscome},
+  title = {Minimal modeling of the extratropical general circulation},
+  journal = {Tellus A: Dynamic Meteorology and Oceanography},
+  volume = {41},
+  number = {4},
+  pages = {292-307},
+  year  = {1989},
+  publisher = {Taylor & Francis},
+  doi = {10.3402/tellusa.v41i4.11842}
 }
 
 @book{K2003,
@@ -303,4 +303,28 @@ doi = {10.3402/tellusa.v41i4.11842}
   author={Kalnay, Eugenia},
   year={2003},
   publisher={Cambridge university press}
+}
+
+@article{Lua,
+  title={Lua—an extensible extension language},
+  author={Ierusalimschy, Roberto and De Figueiredo, Luiz Henrique and Filho, Waldemar Celes},
+  journal={Software: Practice and Experience},
+  volume={26},
+  number={6},
+  pages={635--652},
+  year={1996},
+  publisher={Wiley Online Library},
+  doi={10.1002/(SICI)1097-024X(199606)26:6%3C635::AID-SPE26%3E3.0.CO;2-P}
+}
+
+@article{f2py,
+  title={F2PY: a tool for connecting Fortran and Python programs},
+  author={Peterson, Pearu},
+  journal={International Journal of Computational Science and Engineering},
+  volume={4},
+  number={4},
+  pages={296--305},
+  year={2009},
+  publisher={Inderscience Publishers},
+  doi={10.1504/IJCSE.2009.029165}
 }
