### 2020-11-06 Climdyn/qgs#1 @sadielbartholomew
Add community guidelines to the documentation or README
Towards my review in openjournals/joss-reviews#2597 I have identified that there is not a clear and accessible description for, quoting the review checklist, " third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support".

Please provide this information in a readily-available location, in practice the documentation and/or README file. A few paragraphs could suffice to cover those three components. Thanks.
