### 2020-11-13 #2597 @jodemaey
Hi @sadielbartholomew ,

I have seen that you raised two issues in the qgs repo. Does that mean that your review is complete?

Thank you in advance,

Jonathan
