### 2020-11-25 c6a7dee
Added community guidelines and documented standard model testing Added pytest support Response to issues raised by @sadielbartholomew towards review in openjournals/joss-reviews#2597
