+++ model_test/test_inner_products.py
@@ -1,6 +1,16 @@
 
 # TODO: - Should be rewrited with a string composition function
 
+import sys
+import os
+
+path = os.path.abspath('./')
+base = os.path.basename(path)
+if base == 'model_test':
+    sys.path.extend([os.path.abspath('../')])
+else:
+    sys.path.extend([path])
+
 import unittest
 import numpy as np
 
@@ -14,9 +24,6 @@ real_eps = np.finfo(np.float64).eps
 
 class TestAnalyticInnerProducts(TestBase):
 
-    reference = list()
-    values = list()
-    folder = ""  # 'model_test/'
     filename = 'test_inprod_analytic.ref'
 
     def test_inner_products(self, file=None):
