### 2020-11-25 Climdyn/qgs#1 @jodemaey
Hi,

Thank you for raising that issue.
This issue was also raised by @eviatarbach in his review and is adressed by the commits:

https://github.com/Climdyn/qgs/commit/f87022c6c1f5d5909129fdbba029566f07b1d2bb
https://github.com/Climdyn/qgs/commit/c6a7deefa29f730542a18a206026f269b4c16c2b

in the [joss](https://github.com/Climdyn/qgs/tree/joss) branch. These will be merged into the master when the review process is completed.
