### 2020-11-25 Climdyn/qgs#2 @jodemaey
Hi,

Thank you for raising that issue.
This issue is addressed by the commit:

c6a7dee

in the joss branch. These will be merged into the master when the review process is completed.

The tests can now be executed using `pytest`. `test_base.py` is a base class for the tests and is not used directly.

A recast of the testing system with proper documentation is planned for the v0.3 (or v0.4 at least if we are late).
