### 2020-11-25 #2597 @jodemaey
> Sorry @arfon I still need a bit more time here. Let me promise that by the end of this week (Sunday 29th, UK time) I will finish the review.

Hi @sadielbartholomew ,

Thank you. Meanwhile, I've addressed your two issues on the `qgs` repo.
