### 2020-11-30 #2597 @sadielbartholomew
Hi @jodemaey, I assume that the current state of the proposed JOSS paper content is contained in the `joss` branch rather than `master`? I've just re-generated it via the whedon tool, as per the above comments.

Please in future let us know when you update the paper, since I just spent a significant amount of time reviewing the old version only to see that it has changed quite a bit now, with a new section (that's okay, but going forward it helps for us to know of any state changes).

It's good in the respect that you have, with the updates since the last version generated and linked in this thread (https://github.com/openjournals/joss-reviews/issues/2597#issuecomment-679154289), addressed some aspects that I was going to raise, e.g. interlinking between the references in the text and their corresponding entries in the 'References' section.

However there are some minor aspects of the updated paper itself that I have feedback on. Since these are spread across the `paper.md` , it is clearest if I can refer to specific lines there, so I will use a Pull Request (which I'll open in a moment and link to this Issue) to provide this feedback. Feel free to close it or merge it, as you wish, but consider the contained changeset and comments upon it formal feedback on the paper content (i.e. towards the items listed uner 'Software paper' in the review checklist).

Thanks.
