### 2020-11-30 f12d947
JOSS paper feedback towards openjournals/joss-reviews#2597
