### 2020-11-30 #2597 @jodemaey
Hi @sadielbartholomew ,

Thanks for your suggestions. I will merge them to the joss branch when your review is finished.
