### 2020-11-30 #2597 @jodemaey
Hi @sadielbartholomew ,

I had a look at your commit. Can I merge it as it is?  What about the other items on the checklist about functionalities and documentation: have you considered them already?
