### 2020-11-30 #2597 @sadielbartholomew
> I had a look at your commit. Can I merge it as it is?

Hi @jodemaey, sorry I fell asleep before I put it in as a PR. Please don't merge it yet. Let me open it now since many of my comments are not contained in the commit but I intended to put as comments on said PR. One moment.

> What about the other items on the checklist about functionalities and documentation: have you considered them already?

I am nearly ready to tick these but not quite since I want to be thorough. I'll let you know by the end of the day. Thanks.
