### 2020-11-30 #2597 @sadielbartholomew
Just to note here that I am ticking off items from the review checklist based on the `joss` branch, rather than the `master` branch, as that has been updated in response to feedback and will be the branch used for the state of the codebase for the archive and for the paper content.

Notably, for reviewing the items under 'Documentation' I have built the documentation locally from the `joss` branch (via the documented procedure); I am not referring to the documentation currently hosted online.
