### 2020-11-30 Climdyn/qgs#4 @sadielbartholomew
Distribution installation via a standard package manager
Only manual installation, by cloning (or otherwise obtaining) the source `qgs` codebase, seems to be possible (at least going by what is documented under 'Installation' section and the lack of certain files in the root of the repository). It would be useful for potential users if they were also provided with the option to install the distribution via one or more standard Python installation and dependency management tools such as `pip` or `conda` (the current documented installation instructions only indicate that conda is used to manage the dependencies to create an appropriate environment, rather than that `qgs` itself is available via conda).

Means to enable installation via `pip` and `conda` commands respectively are:

* adding a `setup.py` build script and uploading versions to PyPI; and
* creating a conda recipe and using it to submit the package to a conda channel e.g. anaconda, conda-forge;

where Python have a useful guidance section in their documentation to provide further detail, the [Python Packaging User Guide](https://packaging.python.org/), should that be helpful. I think it is fair to say that it is quite standard nowadays to allow installation at least via `pip`. Is there a reason you have not yet uploaded this to PyPI?

Please note I have noticed this during my review for openjournals/joss-reviews#2597, but it is not a blocker to me recommending acceptance of your paper (as the current installation method works fine), rather a suggestion going forward for improving the package. (You can choose to address it or leave it, and hence close this Issue, as you wish. But I do think it would be really useful to users of `qgs`)
