### 2020-12-01 #2597 @sadielbartholomew
@jodemaey: regarding the installation, the documented approach to installation does work in that it allows the scripts to run successfully, so that is sufficient for the review checklist item. Saying that, I've raised an issue because going forward ideally the package could be made available to install from `PyPI` or some `conda` channel or similar.

I do however have to query a claim in the documentation on that note, namely as stated under the `general_information.html#installation` endpoint of the documentation (from the `joss` branch so built locally):

> The easiest way to install is through Anaconda.

The installation documented there isn't by Anaconda, as conda is only being used to create and activate an environment with the required dependencies. Strictly nothing is being 'installed' at all as no package has been set up, rather we can run `qgs` as documented because the `python` interpreter is being run directly on the set of top-level scripts.

I think the wording of the documentation in this respect needs to be amended to reflect the fact that there is no 'installation' going on per se, else parts such as that sentence are confusing and untrue. An environment is set up with the dependencies, then the scripts, cloned locally, can be run. So for example, the quoted sentence above would in accuracy be something like "The easiest way to run qgs is to use an appropriate environment created through Anaconda" or similar.
