### 2020-12-01 Climdyn/qgs#2 @sadielbartholomew
Thanks @jodemaey, I have ran the tests via the newly-documented means and they all pass:

```pytb
$  pytest
rootdir: /home/sadie/qgs
collected 4 items

model_test/test_aotensor.py .                                            [ 25%]
model_test/test_aotensor_6x6.py .                                        [ 50%]
model_test/test_inner_products.py .                                      [ 75%]
model_test/test_inner_products_6x6.py .                                  [100%]

======================== 4 passed in 418.87s (0:06:58) =========================
```

So I now consider this closed.
