### 2020-12-01 174fb92
Adapted the text in relation to reviewer 2 comment
