### 2020-12-01 Climdyn/qgs#4 @jodemaey
This is presently a research code that may still evolve a lot. So in my view it is not a package.

That said, although not a priority, I could make a package for the v1.0 (if ever) when its structure and code base will allow a broader usage.

As such I will let your issue open. Thanks for the suggestion.
