### 2020-12-01 4c1e0a8
Adapted the text in relation to reviewer 2 comment (cont'd)
