### 2020-12-01 #2597 @jodemaey
@sadielbartholomew : I have responded to your remark about packaging the code on the `qgs` repo issues page: https://github.com/Climdyn/qgs/issues/4
