### 2020-12-01 b19cbad
Put details in footnotes and simplified a sentence. Removed unnecessary mention to f2py.
