### 2020-12-01 #2597 @jodemaey
Hi @sadielbartholomew ,

Please note that new modifications of the paper have been pushed to the joss branch in commit: https://github.com/Climdyn/qgs/commit/b19cbadc8020f8f77f7e451f35d922ff6d1a935c
