### 2020-12-04 bab48df
Added a reference to the DA study of Carrassi et al. Fixed the reference to Tondeur et al.
