+++ documentation/notes/joss_paper/joss.bib
@@ -260,7 +260,8 @@
   title={On temporal scale separation in coupled data assimilation with the ensemble Kalman filter},
   author={Tondeur, Maxime and Carrassi, Alberto and Vannitsem, Stephane and Bocquet, Marc},
   journal={Journal of Statistical Physics},
-  pages={1--25},
+  pages={1161--1185},
+  volume = {179},
   year={2020},
   publisher={Springer},
   doi={10.1007/s10955-020-02525-z}
@@ -339,4 +340,11 @@
   year={2009},
   publisher={Inderscience Publishers},
   doi={10.1504/IJCSE.2009.029165}
+}
+
+@article{CBDGRV2020,
+  title={Data assimilation for chaotic dynamics},
+  author={Carrassi, Alberto and Bocquet, Marc and Demaeyer, Jonathan and Grudzien, Colin and Raanes, Patrick and Vannitsem, Stephane},
+  journal={arXiv preprint arXiv:2010.07063},
+  year={2020}
 }
