### 2020-12-07 #2597 @jodemaey
Hi @sadielbartholomew ,

When can I reasonably expect your review to be final? I need that information to plan my week of work.

Thank you in advance.
