### 2020-12-08 Climdyn/qgs#5 @jodemaey
Be more specific about the installation process in the documentation
Solve the issue raised by the comment https://github.com/openjournals/joss-reviews/issues/2597#issuecomment-736128629 . Modify the Readme and the documentation accordingly.
