### 2020-12-09 #2597 @sadielbartholomew
Thanks for opening up an issue indicating you are going to address my previous comment regarding the installation instructions, @jodemaey. I was going to re-raise that in case you missed it. Otherwise all of my feedback raised so far has been addressed well, bar one small formatting change to the text, described against point (3) below.

I have now completed the remainder of my review and I've outlined the results below (in short I'm happy with everything except for a small number of items I feel still need to be addressed). Thanks for your patience waiting for the rest of the review.

### Final review comments

If the following three points are addressed satisfactorily then I am more than happy to tick the remaining items in the review checklist and moreover to recommend acceptance of the `qgs` paper to JOSS:

1. **Regarding the state of the field**, the review criterion is to 'describe how this software compares to other commonly-used packages' and whilst several similar packages are currently referenced and briefly described under a section of that name, I can't find any *comparison* between qgs and any of those packages in the text. There is some indication of the relation to MAOOAM, but since that is your own package and is included in qgs, I wouldn't count it as a separate package.

    So, what I would like to see towards fulfilling that criterion is some brief comparison between qgs and the listed alternative "easy-to-use idealized atmospheric model[s]" as you introduce them (i.e. q-gcm, pyqg, Isca). A sentence or two on what qgs offers that these libraries do not, or vice versa, would suffice. It may tie in with ground you have already covered in the first two paragraphs of the 'Statement of need' section, so you could perhaps reference them there with some further detail to address this point.

2. **The functionality** of the software is documented really well, which is great. I did feel that the API reference documentation is quite hard to find, though. For example, to see the model configuration classes I would have to go through the menu like so: 'Code Description -> Modules -> Configuration of the model' and then there's a bit of introductory text before the class signature and parameters are listed, so if a reader was viewing on a laptop they might not see it and not realise the classes were covered below.

    In my experience of using and creating software documentation, it is most common to have the API reference as a top-level menu item, under a name such as 'API' or 'reference', where it is evident and easy to access. Therefore I advise you:
    * move the 'Modules' section out to a higher-level in the documentation tree; and/or (but ideally 'and')
    * rename it to one of the names above, as 'Modules' could imply an overview or commentary on the modules rather than the API reference for them.

3. **For the paper text formatting**, please convert `python` to `Python` in the Performance section of the paper text (as raised in Climdyn/qgs#3 but not yet addressed).

Finally, I have **some further comments that are not compulsory for my review acceptance**, but that I think would improve the software or the paper content:

4. I notice the labels on both the bar charts, (a) and (b), in Figure 1  are identical, so to make the visual easier to comprehend I suggest removing the labels on the top chart and moving it so the bottom is joined to the top of the lower chart, or to convert it into a single bar chart with two bars per item on the x-axis.

5. I've noticed that the repository is structured in a way that makes it difficult to search through, namely with the `qgs` functional modules all in separate directories at the root of the repo, instead of being contained within a directory underneath the root which is much more common. This means one can't use a simple `git grep` to search for keywords in the modules to aid with development, as it will pick up words from the Notebook `.ipynb` files (which are big JSON structures with lots of binary code to save the images etc. so clog up the terminal on a search) or the documentation source and build pages, etc., too.

    For that reason, so any contributors can easily work on the core code, personally I'd move all of the module directories under another dedicated directory called `qgs` or `lib` or similar.
