### 2020-12-09 9418899
Added some descriptions to the models proposed in the State of the field section
