### 2020-12-09 e7e7747
Better specification of the installation process
