### 2020-12-09 Climdyn/qgs#5 @jodemaey
Solved through commit e7e77477754bb2cdb36e7b421aff932371e8a895.
