### 2020-12-09 7bc75cf
Moved the code references at the top-level of the doc
