+++ documentation/source/files/references.rst
@@ -1,0 +1,15 @@
+
+References
+----------
+
+Documentation of the classes and functions of the models.
+
+.. toctree::
+   :maxdepth: 2
+
+   technical/configuration
+   technical/inner_products
+   technical/tensors
+   technical/functions
+   technical/integrators
+   technical/misc
