### 2020-12-09 0ff4d83
Add comparison to other software.
