+++ documentation/notes/joss_paper/paper.md
@@ -48,7 +48,7 @@ A future development is planned that will enable the user to specify the basis o
 The model implementation consists of submodules to set up the model's parameters and to compute the tensor that defines the coefficients of the system of ODEs^[More details about the implementation can be found in @DDV2016 and in the *Code Description* section of the included documentation.].
 This tensor is used by the code to compute the tendencies function and its Jacobian matrix. These functions can then be fed to the `qgs` built-in Runge-Kutta integrator or 
 to another integrator implemented by the user. As an example, the usage of the Julia `DifferentialEquations.jl` [@RN2017] integration package through the Python `diffeqpy` [@diffeqpy] package is provided.
-The tangent linear and adjoint models [@K2003] are also available and allow one to conduct easily data assimilation and linear sensitivity analysis experiments.
+The tangent linear and adjoint models [@K2003] are also available and allow one to easily conduct data assimilation and linear sensitivity analysis experiments.
 
 The model implementation uses NumPy [@vCV2011; @O2006] and SciPy [@scipy] for arrays and computations support, as well as Numba [@numba] and sparse [@sparse] to considerably accelerate the tensor products computation used to compute the tendencies.
 
@@ -65,7 +65,7 @@ However, these models are less often considered in literature, despite their dem
 
 `qgs` aims to popularize these systems by providing a fast and easy-to-use Python framework for researchers and teachers to integrate this kind of model. 
 For an efficient handling of the model by users, its documentation is conceived such that its equations and parameters are explained and linked to the code.
-In the future, its development will be done in a modular fashion which enable the connection of the atmosphere to various other subsystems and its usage with built-in and external toolboxes.
+In the future, its development will be done in a modular fashion which enables the connection of the atmosphere to various other subsystems and the use of built-in and external toolboxes.
 
 The choice to use Python was specifically made to facilitate its use in Jupyter [@jupyter] Notebooks and the multiple recent machine learning libraries that are available in this
 language.
@@ -77,36 +77,37 @@ Other software might interest the reader in need of an easy-to-use idealized atm
 * MAOOAM: The Modular Arbitrary-Order Ocean-Atmosphere Model, a coupled ocean-atmosphere model included in `qgs` [@MAOOAM]. 
           Code available in Lua, Fortran and Python.
 * q-gcm: A mid-latitude grid-based ocean-atmosphere model like MAOOAM with two oceanic layers. Code in Fortran, interface in Python [@qgcm].
-* pyqg: A pseudo-spectral Python solver for quasi-geostrophic systems [@pyqg]. Allow one to create and solve multiple-layers quasigeostrophic systems.
+* pyqg: A pseudo-spectral Python solver for quasi-geostrophic systems [@pyqg]. Allows one to create and solve multiple-layer quasigeostrophic systems.
 * Isca: A research General Circulation Model (GCM) to simulate a full planet. Written in Fortran and largely
         configurable with Python scripts, with internal coding changes required for non-standard cases [@Vetal2018; @Isca].
 
-`qgs` distinguishes itself from these other models by a combination of a simplified and configurable geometry, a spectral discretization, an efficient numerical implementation of the ODE system by a sparse tensor multiplication, and the availability of the tangent linear and adjoint models. As such it is very suitable to quickly simulate very long time periods while capturing the most relevant dynamics of the climate system at mid-latitudes.
+`qgs` distinguishes itself from these other models by a combination of a simplified and configurable geometry, a spectral discretization, an efficient numerical implementation of the ODE system by a sparse tensor multiplication, 
+and the availability of the tangent linear and adjoint models. As such it is very suitable to quickly simulate very long time periods while capturing the most relevant dynamics of the climate system at mid-latitudes.
 
[…]
