### 2020-12-09 5723beb
Some rewording of the other models description
