### 2020-12-09 Climdyn/qgs#7 @jodemaey
Move the code modules to a common subdirectory
Feature requested here: https://github.com/openjournals/joss-reviews/issues/2597#issuecomment-741610669 .
