### 2020-12-10 #2597 @jodemaey
Dear all,

Please find below our response to the reviews of the referees (see the posts https://github.com/openjournals/joss-reviews/issues/2597#issuecomment-683972381 and https://github.com/openjournals/joss-reviews/issues/2597#issuecomment-741610669 ).

# Response to referee 1

We thank @eviatarbach for his helpful review. We respond below to the suggested changes:

> qgs provides a modular implementation of a quasi-geostrophic atmospheric model.
> Having previously used the MAOOAM model in research, I was impressed with its
> ease-of-use, highly readable code, the elegance of the tensor-based
> implementation, and built-in tangent linear and adjoint models. qgs builds on
> this by adding two different land components, and making it easy to switch
> between ocean and land coupling. qgs is also written entirely in Python, as
> opposed to MAOOAM that had Fortran and Lua implementations also. I commend the
> authors for their excellent contribution and hope that qgs will become a widely
> used tool in the modelling hierarchy.
>
> Suggested changes:
>
>    1. There are a few missing parts of the documentation, as per the review
>    checklist. The statement of need is provided in the software paper, but it has
>     to be in the documentation as well.
>      * A statement of need: Do the authors clearly state what problems the software is
>       designed to solve and who the target audience is?
>      * Community guidelines:
>      Are there clear guidelines for third parties wishing to 1) Contribute to the
>      software 2) Report issues or problems with the software 3) Seek support

We have added a *statement of need* section in the documentation. We have also added community guidelines on how to
contribute to the development. These additions are contained in the commit f87022c6c1f5d5909129fdbba029566f07b1d2bb in the `joss` branch.

>    2. I would like to see the software paper mention the availability of the
>    tangent linear and adjoint models, as this is a major advantage for data
>    assimilation research, as well as for, e.g., computing the Lyapunov spectrum.
>    Adding an example of the usage of the tangent linear or adjoint models to the
>     documentation would be helpful.

We now mention the tangent linear explicitly in the paper. We have also added a usage example in the documentation.
Thank you for the suggestion. The commits of these additions are cee79b7b53dcfe7d80b535d2badb0b678aeb4761 and 9fd689d8ab2bfcdd548f708e8862c76ae6c706cb .
>    3. This implementation is written in Python instead of Fortran, but uses Numba
>    and sparse matrices to accelerate the computations. I would thus like to see a
>    performance comparison to the Fortran version of MAOOAM, either in the
>    documentation or the software paper (see the performance item on the
>     checklist).

We have now made a benchmark of `qgs` versus MAOOAM lua/fortran code. The results are detailed in a new section of the paper.
   Related commits: dfaeed977de69ef518474db6f2458e6312bf9938 .
> Minor issues:
>    * "Simpy" should be Sympy in the forthcoming developments section.
>    * The link to the notebooks folder works in the GitHub README, but is broken in
>    the readthedocs documentation. Also, the link to the main folder just goes to
>     the main page of the documentation.
>    * "montain" -> "mountain" in the Reinhold and Pierrehumbert (1982) example page
We have also addressed the raised minor issues in the commit bac1d3c67a383b041b686a6c8bc903d159d602b6 .

We thank Eviatar Bach again and hope that the proposed modifications address all the concerns.

# Response to referee 2

We thank @sadielbartholomew for her thorough review.
We address here her final comments, the other ones have already been addressed in the discussion above.

> Thanks for opening up an issue indicating you are going to address my previous
> comment regarding the installation instructions, @jodemaey.
> I was going to re-raise that in case you missed it. Otherwise
> all of my feedback raised so far has been addressed well, bar one small
> formatting change to the text, described against point (3) below.
>
> I have now completed the remainder of my review and I've outlined the results
> below (in short I'm happy with everything except for a small number of items I
> feel still need to be addressed). Thanks for your patience waiting for the rest
> of the review. Final review comments
>
> If the following three points are addressed satisfactorily then I am more than
> happy to tick the remaining items in the review checklist and moreover to
> recommend acceptance of the qgs paper to JOSS:
>
> 1. Regarding the state of the field , the review criterion is to 'describe how this
> software compares to other commonly-used packages' and whilst several similar
> packages are currently referenced and briefly described under a section of that
> name, I can't find any comparison between qgs and any of those packages in the
> text. There is some indication of the relation to MAOOAM, but since that is
> your own package and is included in qgs, I wouldn't count it as a separate
> package. So, what I would like to see towards fulfilling that criterion is some brief
> comparison between qgs and the listed alternative "easy-to-use idealized
> atmospheric model[s]" as you introduce them (i.e. q-gcm, pyqg, Isca). A
> sentence or two on what qgs offers that these libraries do not, or vice versa,
> would suffice. It may tie in with ground you have already covered in the first
> two paragraphs of the 'Statement of need' section, so you could perhaps
> reference them there with some further detail to address this point.

 We have better described the other models in this list and have stressed the main assets of `qgs` in this regard.
   Related commits: 94188990ecbddda0409a98f820c34467962b0eab , 0ff4d83621ffa057722856d39e5141895f7ec32c and 5723bebe6d00dd0354ce70f1f26ae1b81a54af6b .

> 2. The functionality of the software is documented really well, which is great. I
> did feel that the API reference documentation is quite hard to find, though.
> For example, to see the model configuration classes I would have to go through
> the menu like so: 'Code Description -> Modules -> Configuration of the model'
> and then there's a bit of introductory text before the class signature and
> parameters are listed, so if a reader was viewing on a laptop they might not
> see it and not realise the classes were covered below.
> In my experience of using and creating software documentation, it is most common
> to have the API reference as a top-level menu item, under a name such as 'API'
> or 'reference', where it is evident and easy to access. Therefore I advise you:
>   * move the 'Modules' section out to a higher-level in the documentation tree;
>     and/or (but ideally 'and')
>   * rename it to one of the names above, as 'Modules' could imply an overview or
>     commentary on the modules rather than the API reference for them.

We agree that the proposed change will increase the readability of the documentation. Thank you.
We have implemented it in the commit 7bc75cf9f1609d850ea6cc04939c8ce24b8ff912.

> 3. For the paper text formatting , please convert "python" to "Python" in the
> Performance section of the paper text (as raised in [Climdyn/qgs#3](https://github.com/Climdyn/qgs/pull/3) but not yet addressed).

This was done in the commit 4c1e0a85f4997aeb614cd926ed51f5cb06b9306f .

> Finally, I have some further comments that are not compulsory for my review
> acceptance , but that I think would improve the software or the paper content:
> 1. I notice the labels on both the bar charts, (a) and (b), in Figure 1 are
> identical, so to make the visual easier to comprehend I suggest removing the
> labels on the top chart and moving it so the bottom is joined to the top of the
> lower chart, or to convert it into a single bar chart with two bars per item on
> the x-axis.

We do not think that this modification would make the figure easier to read. Creating a single plot could indeed be a good solution,
   but the data of the two actual panels are too different and that would decrease the readability of the panel *(a)* data.

> 2. I've noticed that the repository is structured in a way that makes it difficult
> to search through, namely with the qgs functional modules all in separate
> directories at the root of the repo, instead of being contained within a
> directory underneath the root which is much more common. This means one can't
> use a simple git grep to search for keywords in the modules to aid with
> development, as it will pick up words from the Notebook .ipynb files (which are
> big JSON structures with lots of binary code to save the images etc. so clog up
> the terminal on a search) or the documentation source and build pages, etc.,
> too.
> For that reason, so any contributors can easily work on the core code,
> personally I'd move all of the module directories under another dedicated
> directory called qgs or lib or similar.
>

Thank you for this suggestion. We could consider changing the structure of the repository to indeed simplify it along the line that you propose.
However, we do not want to make that change now, since some consequent refactorings are planned for the v0.3. We need to think about it, and as such
we have opened the issue https://github.com/Climdyn/qgs/issues/7 to discuss and come back to it later.

We thank Sadie Bartholomew again for her suggestions.

# General comment

Various small language changes and consistency have also been applied.
Please see the commits f8d8e41c2a6e2b676ee242a8f55e791cc1de303e and
a318efbafdd4324ffb567ef70171ab222aae80d9 for a recap.
