### 2020-12-10 #2597 @sadielbartholomew
Thanks @jodemaey for the considered response. **I have inspected your review response comment and the referenced commits and I am now satisfied that all of the review checklist items are covered well by the `qgs` code and the proposed paper and am delighted to recommend that the `qgs` paper be accepted.**

Just a few closing comments, only the first of which calls for any action (ideally it could be addressed before the paper is released should everyone else is happy for it to be accepted).

I rebuilt the documentation locally to check on the restructure for the API reference and there is now a top-level 'References' section which addresses the feedback at hand, but there seems to be a setup issue since many (but not all) of the pages, notably:

* Inner Products (living at `<path>/documentation/build/html/files/technical/inner_products.html`)
* Tensors module (`.../tensors.html`)
* Functions module (etc.)
* Integration modules

all seem to be blank, whereas in the documentation currently live on `qgs.readthedocs.io` those pages appear to be populated with the reference information. So that should be addressed to ensure the latest documentation retains all the reference information.

> We do not think that this modification would make the figure easier to read. Creating a single plot could indeed be a good solution,
but the data of the two actual panels are too different and that would decrease the readability of the panel (a) data.

That is fair enough. The plots are certainly readable as they are.

> Thank you for this suggestion. We could consider changing the structure of the repository to indeed simplify it along the line that you propose.
However, we do not want to make that change now, since some consequent refactorings are planned for the v0.3. We need to think about it, and as such
we have opened the issue Climdyn/qgs#7 to discuss and come back to it later.

I'm glad to hear you are considering this, of course address it at a time which is most appropriate in your planned roadmap.
