### 2020-12-11 #2597 @jodemaey
Hi @sadielbartholomew ,

> I rebuilt the documentation locally to check on the restructure for the API reference and there is now a top-level 'References' section which addresses the feedback at hand, but there seems to be a setup issue since many (but not all) of the pages, notably:
>
>     * Inner Products (living at `<path>/documentation/build/html/files/technical/inner_products.html`)
>
>     * Tensors module (`.../tensors.html`)
>
>     * Functions module (etc.)
>
>     * Integration modules
>
>
> all seem to be blank, whereas in the documentation currently live on `qgs.readthedocs.io` those pages appear to be populated with the reference information. So that should be addressed to ensure the latest documentation retains all the reference information.
>

I cannot reproduce your problem, on my system the documentation pages that you mention are fine.
Could you first perform a `make clean` before doing a `make html` again?
