### 2020-12-11 #2597 @sadielbartholomew
Hi @jodemaey, thanks for confirming you can't recreate it. It was indeed my fault, sorry, I didn't realise but I built the local documentation last time with a different environment than the `qgs` conda one I used for reviewing, such that I did not have the right dependencies in my environment for Sphinx to use to parse the code to generate the API reference. I just tried it with the right environment activated and everything was there as it should be.

So everything is great now as far as I am concerned. Thanks again.
