### 2020-12-12 #2597 @arfon
@jodemaey - please ignore these suggestions from Whedon, it looks like these DOIs are fine (although the URLs are somewhat unusual).

```
INVALID DOIs

- 10.1175/1520-0469(1987)044%3C3282:TOMEAW%3E2.0.CO;2 is INVALID
- 10.1175/1520-0493(1982)110%3C1105:DOWRQS%3E2.0.CO;2 is INVALID
- 10.1002/(SICI)1097-024X(199606)26:6%3C635::AID-SPE26%3E3.0.CO;2-P is INVALID
```

Please check though whether these DOIs are correct, and if so, please add them to your BibTeX:

```
MISSING DOIs

- 10.2142/biophys.48.324 may be a valid DOI for title: Isca
- 10.1017/cbo9780511617652.004 may be a valid DOI for title: Predictability: A problem partly solved
```
