### 2020-12-13 ec6b48c
Formatting fix for paper.
