### 2020-12-13 #2597 @arfon
@jodemaey - I also found this small fix for a formatting issue with one reference: https://github.com/Climdyn/qgs/pull/8
