### 2020-12-13 #2597 @arfon
@jodemaey - after you've fixed up the DOIs and merged https://github.com/Climdyn/qgs/pull/8, could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
