### 2020-12-14 #2597 @jodemaey
>
> Please check though whether these DOIs are correct, and if so, please add them to your BibTeX:
>
> ```
> MISSING DOIs
>
> - 10.2142/biophys.48.324 may be a valid DOI for title: Isca
> - 10.1017/cbo9780511617652.004 may be a valid DOI for title: Predictability: A problem partly solved
> ```

Hi @arfon ,

Thank you for your remark. It was already raised during the pre-review. Here was my answer to it: https://github.com/openjournals/joss-reviews/issues/2549#issuecomment-669902829 .

Basically the DOI Whedon proposes for Isca is wrong, but the description paper of Isca is correctly cited in our paper.
And the DOI for the Lorenz paper (*Predictability: A problem partly solved*) actually point to a re-edition but I want to cite the original one here. There is no DOI for the original one.
