### 2020-12-14 #2597 @jodemaey
> And the DOI for the Lorenz paper (_Predictability: A problem partly solved_) actually point to a re-edition but I want to cite the original one here. There is no DOI for the original one.

Maybe I could cite the original article but then join to its citation the re-edition DOI so that people still get access to an electronic version.
