### 2020-12-17 #2597 @arfon
> Maybe I could cite the original article but then join to it the citation the DOI of the re-edition so that people still get a link to an electronic version.

Yes, that sounds like a reasonable solution.
