### 2020-12-17 #2597 @jodemaey
Hi @arfon ,

Sorry but after discussing with my co-authors, we don't think it is a good idea to put the DOI of one citation corresponding to another one.
Again we prefer to cite the original Lorenz 96 article since it is the one that gave its name to the model being referred to. It is also the seminal reference that is usually cited in our field (1163 times according to google).
It's a seminar proceeding that was never officially published and thus never got a DOI, however it is quite easy to find it on many place on the web with a quick google search.

If you agree with this, I will proceed and make the release.

Jonathan
