### 2020-12-20 27acae0
Changed the version number to v0.2.1
