### 2020-12-20 1f16589
JOSS article commit:
* manuscript
* modification and improvement of the documentation and README
* Added tests for a higher dimensional (6x6) model version
