+++ documentation/notes/joss_paper/paper.md
@@ -1,0 +1,117 @@
+---
+title: 'qgs: A flexible Python framework of reduced-order multiscale climate models'
+tags:
+  - Python
+  - Numba
+  - Idealized atmospheric model
+  - Coupled model
+  - Mid-latitude climate variability
+authors:
+  - name: Jonathan Demaeyer
+    orcid: 0000-0002-5098-404X 
+    affiliation: 1
+  - name: Lesley De Cruz
+    orcid: 0000-0003-4458-8953
+    affiliation: 1
+  - name: Stéphane Vannitsem 
+    orcid: 0000-0002-1734-1042
+    affiliation: 1
+affiliations:
+ - name: Royal Meteorological Institute of Belgium, Avenue Circulaire, 3, 1180 Brussels, Belgium
+   index: 1
+
+date: 7 August 2020
+bibliography: joss.bib
+
+---
+
+# Summary
+
+`qgs` is a a Python implementation of a set of idealized reduced-order models representing atmospheric mid-latitude variability. 
+It consists of a two-layer *q*uasi-*g*eostrophic *s*pectral (`qgs`) model of the atmosphere on a beta-plane, coupled either to a simple land surface or to a shallow-water ocean. 
+The model's dynamical fields include the atmospheric and oceanic streamfunction and temperature fields, and the land temperature field.
+
+* In the case where it is coupled to an ocean, it reproduces the Modular Arbitrary-Order Ocean-Atmosphere Model (MAOOAM), described in @DDV2016. In @VDDG2015, a 36-variable configuration of this model was shown to reproduce a 
+low-frequency variability (LFV) typical of the coupled ocean-atmosphere system. 
+This coupling consists in both mechanical and heat exchange interactions between the two components. 
+The model has already been used in different contexts, in particular for data assimilation [@PBBCDSY2019; @TCVB2020], 
+and predictability studies [@VSD2019; @VD2020]
+* In the case of a land surface coupling, it emulates the model proposed in @RP1982 and @CT1987 with a simple thermal relaxation toward a climatological temperature and a mechanical coupling due to the 
+friction between the land and the atmosphere. It can also emulate the model proposed in @LHHBD2018, with mechanical coupling and heat exchange. In addition, the number of dynamical spectral modes can be configured by the user, as is the case for the MAOOAM model.
+
+In the `qgs` framework, the partial differential equations (PDEs) that govern the time evolution of its fields are projected on a basis of functions defined on its 
+spatial domain. 
+This kind of decomposition transforms the PDEs into a set of ordinary differential equations (ODEs) which can then be solved with the usual integration techniques.
+Presently in `qgs`, the functions of the basis are chosen amongst the orthogonal Fourier modes compatible with the boundary conditions of each subcomponent of the system, namely the atmosphere, and the ocean or the land surface coupled to it. 
+A future development is planned that will enable the user to specify the basis of functions for each component, depending on the required boundary conditions.
+
+The model implementation consists of submodules to set up the model's parameters and to compute the tensor that defines the coefficients of the system of ODEs^[More details about the implementation can be found in @DDV2016 and in the *Code Description* section of the included documentation.].
+This tensor is used by the code to compute the tendencies function and its Jacobian matrix. These functions can then be fed to the `qgs` built-in Runge-Kutta integrator or 
+to another integrator implemented by the user. As an example, the usage of the Julia `DifferentialEquations.jl` [@RN2017] integration package through the Python `diffeqpy` [@diffeqpy] package is provided.
+The tangent linear and adjoint models [@K2003] are also available and allow one to easily conduct data assimilation and linear sensitivity analysis experiments.
+
[…]
