+++ documentation/source/execute_examples.sh
@@ -1,8 +1,0 @@
-#!/usr/bin/env bash
-
-cd source/files/examples
-
-for file in ./*.ipynb; do
-    jupyter nbconvert --execute --inplace "$file"
-done
-
