+++ documentation/execute_examples.sh
@@ -3,5 +3,6 @@
 cd source/files/examples || exit
 
 for file in ./*.ipynb; do
-    jupyter nbconvert --to rst "$file"
+    jupyter nbconvert --execute --inplace --ExecutePreprocessor.timeout=600 "$file"
 done
+
