### 2020-12-20 #2597 @jodemaey
Hi @arfon ,

I've just merged the JOSS branch in master and released a new version [v0.2.1](https://github.com/Climdyn/qgs/releases/tag/v0.2.1) containing the paper and the modifications. Zenodi has automatically generated a new DOI: 10.5281/zenodo.4368844 that is already linked on the main page. Do I need to generate a specific archive for the article as you mentioned?

Also, please note that the online doc is currently not updated due to a [bug](https://github.com/executablebooks/jupyter-book/issues/1137#issuecomment-743993738) .
